General 404 handling

+++ public/index.php
@@ -1,38 +1,52 @@
 <?php
 
-ini_set('error_reporting', E_ALL);
-ini_set('display_errors', 1);
-
 require_once __DIR__ . '/../class-loader.php';
 
 $app = new Slim\Slim();
 
-switch($app->request()->getMethod()) {
-    case 'GET':
-        $app->get('/:resource(/:id)', function($resource, $id = null) {
-            $resource = \App\Resource::load($resource);
-            $resource->get($id);
-        });
-        break;
-    case 'POST':
-        $app->post('/:resource', function($resource) {
-            $resource = \App\Resource::load($resource);
-            $resource->post();
-        });
-        break;
-    case 'PUT':
-        $app->put('/:resource(/:id)', function($resource, $id = null) {
-            $resource = \App\Resource::load($resource);
-            $resource->put($id);
-        });
-        break;
-    case 'DELETE':
-        $app->delete('/:resource(/:id)', function($resource, $id = null) {
-            $resource = \App\Resource::load($resource);
-            $resource->delete($id);
-        });
-        break;
+// Get
+$app->get('/:resource(/:id)', function($resource, $id = null) {
+    $resource = \App\Resource::load($resource);
+    if ($resource === null) {
+        \App\Resource::response(\App\Resource::STATUS_NOT_FOUND);
+    } else {
+        $resource->get($id);
+    }
+});
 
-}
+// Post
+$app->post('/:resource', function($resource) {
+    $resource = \App\Resource::load($resource);
+    if ($resource === null) {
+        \App\Resource::response(\App\Resource::STATUS_NOT_FOUND);
+    } else {
+        $resource->post();
+    }
+});
+
+// Put
+$app->put('/:resource/:id', function($resource, $id = null) {
+    $resource = \App\Resource::load($resource);
+    if ($resource === null) {
+        \App\Resource::response(\App\Resource::STATUS_NOT_FOUND);
+    } else {
+        $resource->put($id);
+    }
+});
+
+// Delete
+$app->delete('/:resource/:id', function($resource, $id = null) {
+    $resource = \App\Resource::load($resource);
+    if ($resource === null) {
+        \App\Resource::response(\App\Resource::STATUS_NOT_FOUND);
+    } else {
+        $resource->delete($id);
+    }
+});
+
+// Not found
+$app->notFound(function() {
+    \App\Resource::response(\App\Resource::STATUS_NOT_FOUND);
+});
 
 $app->run();
